requirements.txt
pandas
numpy
plotly
jinja2
flask
dash
matplotlib
seaborn
scikit-learn
openpyxl
python-pptx
